Rename conf to config
