+++ hugophotoswipe/config.py
